reacomode redux para que sea mas entendible y reacomode los componentes que lo usan
