+++ frontend/src/Components/Detail/Detail.jsx
@@ -3,13 +3,14 @@ import { useEffect } from "react";
 import * as ReactRedux from "react-redux";
 import { useParams } from "react-router-dom";
 import {
-  addFav,
-  addToCart, 
   cleanDetail,
   getDetail,
+} from "../../redux/actions/ProdDetail/ProdDetailAction";
+import {
+  addFav,
   removeFav,
-  
-} from "../../redux/actions";
+} from "../../redux/actions/Favorites/FavoritesAction";
+import { addToCart } from "../../redux/actions/Cart/CartAction";
 import Loading from "../Loading/Loading";
 import img from "../Img/PG0.png";
 import style from "./Details.module.css";
@@ -30,18 +31,22 @@ function Detail() {
     };
   }, [dispatch, id]);
 
-  const product = ReactRedux.useSelector((state) => state.prodDetail);
-  const favorites = ReactRedux.useSelector((state) => state.favorites);
-  const cart = ReactRedux.useSelector((state) => state.cartproduct);
+  const { prodDetail } = ReactRedux.useSelector(
+    (state) => state.prodDetailReducer
+  );
+  const { favorites } = ReactRedux.useSelector(
+    (state) => state.favoriteReducer
+  );
+  const { cartproduct } = ReactRedux.useSelector((state) => state.cartReducer);
 
   const addFav1 = () => {
-    if (product.name) {
-      dispatch(addFav(product));
+    if (prodDetail.name) {
+      dispatch(addFav(prodDetail));
     }
   };
 
   const removeFav1 = (id) => {
-    if (product.id) {
+    if (prodDetail.id) {
       dispatch(removeFav(id));
     }
   };
@@ -54,61 +59,61 @@ function Detail() {
   };
 
   const isOnCart = () => {
-    if (cart?.length) {
-      return cart.find((p) => p.id === product.id) ? true : false;
+    if (cartproduct?.length) {
+      return cartproduct.find((p) => p.id === prodDetail.id) ? true : false;
     }
     return false;
   };
 
-   const addCart = () => {
+  const addCart = () => {
     if (isOnCart()) {
-       return swal({
-         title: "Cuidado",
-         text: "El producto ya esta en el carrito",
-         icon: "error",
-         button: "Aceptar",
-         timer: "2500"
-       });
-     }
-     dispatch(addToCart(product));
-     swal({
-       title: "Exito",
-       text: `${product.name} agregado al carrito`,
-       icon: "success",
-       button: "Aceptar",
-       timer: "2500"
-     });
-   };
+      return swal({
+        title: "Cuidado",
+        text: "El producto ya esta en el carrito",
+        icon: "error",
+        button: "Aceptar",
+        timer: "2500",
+      });
+    }
+    dispatch(addToCart(prodDetail));
+    swal({
+      title: "Exito",
+      text: `${prodDetail.name} agregado al carrito`,
+      icon: "success",
+      button: "Aceptar",
+      timer: "2500",
+    });
+  };
 
   return (
     <div>
-      {product.name ? (
+      {prodDetail.name ? (
         <div className={style.style}>
-          <h1 className={style.title}>{product.name}</h1>
+          <h1 className={style.title}>{prodDetail.name}</h1>
 
           <img
             className={style.img}
-            alt={product.name}
-            src={product.image ? product.image : img}
+            alt={prodDetail.name}
+            src={prodDetail.image ? prodDetail.image : img}
           />
           <div className={style.details}>
-            <p className={style.details}>{product.longDescription}</p>
-            {product.categories.map((cat) => (
+            <p className={style.details}>{prodDetail.longDescription}</p>
+            {prodDetail.categories.map((cat) => (
               <p key={cat.id}>{cat.name}</p>
             ))}
           </div>
 
           <section>
-
-            <h3 className={style.stock}>Stock disponible:{product.stock} unidades</h3>
+            <h3 className={style.stock}>
+              Stock disponible:{prodDetail.stock} unidades
+            </h3>
           </section>
           <section>
-
-            <h3 className={style.precio}>Precio: ${product.price}</h3>
+            <h3 className={style.precio}>Precio: ${prodDetail.price}</h3>
             {!isAuthenticated ? (
               <></>
-            ) : prodIsFav(product.id) ? (
-              <button onClick={() => removeFav1(product.id)}>
+            ) : prodIsFav(prodDetail.id) ? (
+              <button onClick={() => removeFav1(prodDetail.id)}>
                 <FcLike />
               </button>
             ) : (
@@ -116,7 +121,12 @@ function Detail() {
                 <AiOutlineHeart />
               </button>
             )}
-            { <button className={style.button} onClick={addCart}> Agregar al Carro </button> }
+            {
+              <button className={style.button} onClick={addCart}>
+                {" "}
+                Agregar al Carro{" "}
+              </button>
+            }
             <button className={style.button1}> Comprar </button>
           </section>
         </div>
