--- frontend/src/redux/actions/ActionTypes.js
@@ -1,19 +1,0 @@
-export const GET_NAME_PRODUCT = "GET_NAME_PRODUCT";
-export const GET_PRODUCTS = "GET_PRODUCTS";
-export const ORDER_ALPHABETICAL = "ORDER_ALPHABETICAL";
-export const ORDER_PRICE = "ORDER_PRICE";
-export const GET_DETAIL = "GET_DETAIL";
-export const GET_PRODUCT_BY_CATEGORY = "GET_PRODUCT_BY_CATEGORY";
-export const GET_CATEGORIES = "GET_CATEGORIES";
-export const CLEAN_DETAIL = "CLEAN_DETAIL";
-export const FILTER_BY = "FILTER_BY";
-export const SET_PAGE = "SET_PAGE";
-export const WEEK_PROD = "WEEK_PROD";
-export const FILTER_BY2 = "FILTER_BY2";
-export const ADD_TO_CART = "ADD_TO_CART";
-export const ORDER_PRODUCT = "ORDER_PRODUCT";
-export const REMOVE_ONE_FROM_CART = "REMOVE_ONE_FROM_CART";
-export const REMOVE_ALL_FROM_CART = "REMOVE_ALL_FROM_CART";
-export const CLEAR_CART = "CLEAR_CART";
-export const ADD_FAV = "ADD_FAV";
-export const REMOVE_FAV = "REMOVE_FAV";
