+++ frontend/src/redux/actions/Cart/ActionTypes.js
@@ -1,0 +1,5 @@
+export const ADD_TO_CART = "ADD_TO_CART";
+export const ORDER_PRODUCT = "ORDER_PRODUCT";
+export const REMOVE_ONE_FROM_CART = "REMOVE_ONE_FROM_CART";
+export const REMOVE_ALL_FROM_CART = "REMOVE_ALL_FROM_CART";
+// export const CLEAR_CART = "CLEAR_CART";
