+++ frontend/src/redux/actions/Categories/ActionTypes.js
@@ -1,0 +1,2 @@
+export const GET_PRODUCT_BY_CATEGORY = "GET_PRODUCT_BY_CATEGORY";
+export const GET_CATEGORIES = "GET_CATEGORIES";
