+++ frontend/src/redux/actions/Favorites/ActionTypes.js
@@ -1,0 +1,2 @@
+export const ADD_FAV = "ADD_FAV";
+export const REMOVE_FAV = "REMOVE_FAV";
