+++ frontend/src/redux/actions/Page/ActionTypes.js
@@ -1,0 +1,1 @@
+export const SET_PAGE = "SET_PAGE";
