+++ frontend/src/redux/actions/Page/PageAction.js
@@ -1,0 +1,8 @@
+import { SET_PAGE } from "./ActionTypes";
+
+export const setPageAct = (page) => {
+    return {
+      type: SET_PAGE,
+      payload: page,
+    };
+  };
