+++ frontend/src/redux/actions/ProdDetail/ActionTypes.js
@@ -1,0 +1,2 @@
+export const GET_DETAIL = "GET_DETAIL";
+export const CLEAN_DETAIL = "CLEAN_DETAIL";
