+++ frontend/src/redux/actions/ProdDetail/ProdDetailAction.js
@@ -1,0 +1,26 @@
+import axios from "axios";
+import { CLEAN_DETAIL, GET_DETAIL } from "./ActionTypes";
+
+export const getDetail = (id) => {
+  return async (dispatch) => {
+    try {
+      const product = await axios.get(
+        `https://pf-api-04.up.railway.app/products/${id}`
+      );
+
+      return dispatch({
+        type: GET_DETAIL,
+        payload: product.data,
+      });
+    } catch (error) {
+      console.log(error);
+    }
+  };
+};
+
+export const cleanDetail = () => {
+  return {
+    type: CLEAN_DETAIL,
+    payload: {},
+  };
+};
