+++ frontend/src/redux/actions/Products/ActionTypes.js
@@ -1,0 +1,6 @@
+export const GET_NAME_PRODUCT = "GET_NAME_PRODUCT";
+export const GET_PRODUCTS = "GET_PRODUCTS";
+export const ORDER_ALPHABETICAL = "ORDER_ALPHABETICAL";
+export const ORDER_PRICE = "ORDER_PRICE";
+export const FILTER_BY = "FILTER_BY";
+export const FILTER_BY2 = "FILTER_BY2";
