+++ frontend/src/redux/store/index.js
@@ -1,11 +1,9 @@
-import { createStore, applyMiddleware } from 'redux';
-import { composeWithDevTools } from 'redux-devtools-extension';
-import reducer from '../reducer/index';
-import thunk from 'redux-thunk';
+import { createStore, applyMiddleware } from "redux";
+import { composeWithDevTools } from "redux-devtools-extension";
+import { rootReducer } from "../reducer/index";
+import thunk from "redux-thunk";
 
 export const store = createStore(
-    reducer,
-    composeWithDevTools(
-        applyMiddleware(thunk)
-    )
+  rootReducer,
+  composeWithDevTools(applyMiddleware(thunk))
 );
